Added feather shooting to crow

+++ GameJam2022Unity/Assets/Scripts/Enemy AI/Boar.cs
@@ -4,7 +4,7 @@ using UnityEngine;
 
 public class Boar : MonoBehaviour
 {
-    [SerializeField] public int attackDistance = 3;
+    [SerializeField] public float attackDistance = 3.0f;
     [SerializeField] public GameObject chargeAttack;
     [SerializeField] public float noticePlayerJumpForce = 100.0f;
     [SerializeField] public float m_MovementSmoothing = 0.05f;
